packages/forge-gauge/pyproject.toml
[project]
name = "forge-gauge"
version = "2.2.0"
description = "Container vulnerability scanning and image matching tool"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [
    { name = "Chainguard Field Engineering" }
]
dependencies = [
    "forge-core",
    "anthropic>=0.40.0,<1.0.0",
    "markdown>=3.0.0,<4.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "requests>=2.32.0,<3.0.0",
    "xlsxwriter>=3.2.0,<4.0.0",
]

[project.entry-points."forge.plugins"]
gauge = "forge_gauge:create_plugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge_gauge"]

[tool.uv.sources]
forge-core = { workspace = true }

[tool.mypy]
python_version = "3.12"
strict = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
